initial commit
Some algorithmic measures have been taken out to explore and dictate the behavior of such naval propulsion plants and their processes, and also to predict the GT Turbine decay state coefficient using different machine learning algorithms. Since, the dataset along with its features elaborates its continuous behavior, the algorithms used in this scenario are the regression algorithms, such as XGBoost regressor, Multiple Linear Regression, and RandomForest regressor. Usually, the dataset is full of noise and redundancy, for that some preprocessing techniques have been applied for the preparation of data. The models embedded with algorithms have gone through some  hyperparameter tuning so as to express their best performance with the tuned and appropriate parameters.